feat: add option geekdocDarkModeCode to enforce dark codeblocks
