+++ layouts/partials/site-header.html
@@ -31,7 +31,7 @@
         {{ end }}
 
 
-        <span id="gdoc-dark-mode">
+        <span id="gdoc-color-theme">
           <svg class="gdoc-icon gdoc_brightness_dark">
             <title>{{ i18n "button_toggle_dark" }}</title>
             <use xlink:href="#gdoc_brightness_dark"></use>
