+++ src/sass/_base.scss
@@ -1,23 +1,31 @@
 :root,
-:root[color-mode="light"] {
+:root[color-theme="light"] {
   --code-max-height: none;
 
-  @include light_mode;
+  @include color_theme_light;
+  @include code_theme_light;
 }
 
 @media (prefers-color-scheme: light) {
   :root {
-    @include light_mode;
+    @include color_theme_light;
+    @include code_theme_light;
   }
 }
 
-:root[color-mode="dark"] {
-  @include dark_mode;
+:root[color-theme="dark"] {
+  @include color_theme_dark;
+  @include code_theme_dark;
+}
+
+:root[code-theme="dark"] {
+  @include code_theme_dark;
 }
 
 @media (prefers-color-scheme: dark) {
   :root {
-    @include dark_mode;
+    @include color_theme_dark;
+    @include code_theme_dark;
   }
 }
 
@@ -27,7 +35,7 @@ html {
   scroll-behavior: smooth;
 
   &.color-toggle-hidden {
-    #gdoc-dark-mode {
+    #gdoc-color-theme {
       .gdoc_brightness_auto,
       .gdoc_brightness_dark,
       .gdoc_brightness_light {
@@ -37,7 +45,7 @@ html {
   }
 
   &.color-toggle-light {
-    #gdoc-dark-mode {
+    #gdoc-color-theme {
       .gdoc_brightness_light {
         display: inline-block;
       }
@@ -49,7 +57,7 @@ html {
   }
 
   &.color-toggle-dark {
-    #gdoc-dark-mode {
+    #gdoc-color-theme {
       .gdoc_brightness_dark {
         display: inline-block;
       }
@@ -61,7 +69,7 @@ html {
   }
 
   &.color-toggle-auto {
-    #gdoc-dark-mode {
+    #gdoc-color-theme {
       .gdoc_brightness_light {
         display: none;
       }
@@ -131,7 +139,7 @@ img {
   vertical-align: middle;
 }
 
-#gdoc-dark-mode {
+#gdoc-color-theme {
   cursor: pointer;
 }
